+++ src/core/tnlField2D.h
@@ -30,8 +30,8 @@ template< class T > class tnlField2D : public tnlLongVector< T >
    : tnlLongVector< T >( 0 )
    { };
 
-   tnlField2D( long int _x_size,
-             long int _y_size )
+   tnlField2D( int _x_size,
+               int _y_size )
    : tnlLongVector< T >( _x_size * _y_size ),
      x_size( _x_size ), y_size( _y_size )
    { };
@@ -47,18 +47,18 @@ template< class T > class tnlField2D : public tnlLongVector< T >
       return tnlString( "tnlField2D< " ) + tnlString( GetParameterType( t ) ) + tnlString( " >" );
    };
 
-   long int GetXSize() const
+   int GetXSize() const
    {
       return x_size;
    };
 
-   long int GetYSize() const
+   int GetYSize() const
    {
       return y_size;
    };
 
-   bool SetNewDimensions( long int new_x_size,
-                          long int new_y_size )
+   bool SetNewDimensions( int new_x_size,
+                          int new_y_size )
    {
       x_size = new_x_size;
       y_size = new_y_size;
@@ -70,26 +70,26 @@ template< class T > class tnlField2D : public tnlLongVector< T >
       return SetNewDimensions( f. GetXSize(), f. GetYSize() );
    };
 
-   void SetSharedData( T* _data, const long int _x_size, const long int _y_size )
+   void SetSharedData( T* _data, const int _x_size, const int _y_size )
    {
       tnlLongVector< T > :: SetSharedData( _data, _x_size * _y_size );
       x_size = _x_size;
       y_size = _y_size;
    };
 
-   const T& operator() ( long int i, long int j ) const
+   const T& operator() ( int i, int j ) const
    {
       assert( i < x_size && j < y_size && i >= 0 && j >= 0 );
       return tnlLongVector< T > :: data[ i * y_size + j ];
    };
 
-   T& operator() ( long int i, long int j )
+   T& operator() ( int i, int j )
    {
       assert( i < x_size && j < y_size && i >= 0 && j >= 0 );
       return tnlLongVector< T > :: data[ i * y_size + j ];
    };
 
-   long int GetLongVectorIndex( long int i, long int j ) const
+   int GetLongVectorIndex( int i, int j ) const
    {
       assert( i >= 0 && j >= 0 );
       assert( i < x_size && j < y_size );
@@ -100,8 +100,8 @@ template< class T > class tnlField2D : public tnlLongVector< T >
    bool Save( ostream& file ) const
    {
       if( ! tnlLongVector< T > :: Save( file ) ) return false;
-      file. write( ( char* ) &x_size, sizeof( long int ) );
-      file. write( ( char* ) &y_size, sizeof( long int ) );
+      file. write( ( char* ) &x_size, sizeof( int ) );
+      file. write( ( char* ) &y_size, sizeof( int ) );
       if( file. bad() ) return false;
       return true;
    };
@@ -110,15 +110,15 @@ template< class T > class tnlField2D : public tnlLongVector< T >
    bool Load( istream& file )
    {
       if( ! tnlLongVector< T > :: Load( file ) ) return false;
-      file. read( ( char* ) &x_size, sizeof( long int ) );
-      file. read( ( char* ) &y_size, sizeof( long int ) );
+      file. read( ( char* ) &x_size, sizeof( int ) );
+      file. read( ( char* ) &y_size, sizeof( int ) );
       if( file. bad() ) return false;
       return true;
    };   
    
    protected:
 
-   long int x_size, y_size;
+   int x_size, y_size;
 };
 
 // Explicit instatiation
